Move all integration code to the new project
