--- .github/workflows/integration-tests.yml
@@ -1,65 +1,0 @@
-name: "Integration Tests"
-
-on:
-  schedule:
-    # Run daily at 03:00 UTC
-    - cron: '0 3 * * *'
-  workflow_dispatch:
-    # Allow manual triggering
-  push:
-    branches:
-      - master
-    paths:
-      - 'WalletWasabi.IntegrationTests/**'
-
-jobs:
-  integration-tests:
-    runs-on: ubuntu-latest
-    timeout-minutes: 60
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup .NET
-        uses: actions/setup-dotnet@v4
-        with:
-          dotnet-version: '10.0.x'
-
-      - name: Restore dependencies
-        run: dotnet restore WalletWasabi.IntegrationTests/WalletWasabi.IntegrationTests.csproj
-
-      - name: Build
-        run: dotnet build WalletWasabi.IntegrationTests/WalletWasabi.IntegrationTests.csproj --no-restore --configuration Release
-
-      - name: Make bitcoind executable
-        run: |
-          chmod +x WalletWasabi.Tests/BundledApps/Binaries/linux-x64/bitcoind
-
-      - name: Run Integration Tests
-        run: |
-          dotnet test WalletWasabi.IntegrationTests/WalletWasabi.IntegrationTests.csproj \
-            --no-build \
-            --configuration Release \
-            --logger "console;verbosity=detailed" \
-            --logger "trx;LogFileName=integration-test-results.trx" \
-            --results-directory ./TestResults
-        timeout-minutes: 45
-        env:
-          DOTNET_CLI_TELEMETRY_OPTOUT: 1
-
-      - name: Upload Test Results
-        uses: actions/upload-artifact@v4
-        if: always()
-        with:
-          name: integration-test-results
-          path: ./TestResults/*.trx
-          retention-days: 30
-
-      - name: Publish Test Results
-        uses: dorny/test-reporter@v1
-        if: always()
-        with:
-          name: Integration Tests Report
-          path: './TestResults/*.trx'
-          reporter: dotnet-trx
-          fail-on-error: true
