+++ WalletWasabi.IntegrationTests/BitcoinCore/Configuration/Whitening/WhiteBind.cs
@@ -1,0 +1,10 @@
+using System.Diagnostics.CodeAnalysis;
+using NBitcoin;
+
+namespace WalletWasabi.IntegrationTests.BitcoinCore.Configuration.Whitening;
+
+public class WhiteBind : WhiteEntry
+{
+	public static bool TryParse(string value, Network network, [NotNullWhen(true)] out WhiteBind? white)
+		=> TryParse<WhiteBind>(value, network, out white);
+}
