+++ WalletWasabi.IntegrationTests/BitcoinCore/Endpointing/EndPointStrategyType.cs
@@ -1,0 +1,8 @@
+namespace WalletWasabi.IntegrationTests.BitcoinCore.Endpointing;
+
+public enum EndPointStrategyType
+{
+	Default,
+	Custom,
+	Random
+}
