+++ WalletWasabi.IntegrationTests/BitcoinCore/Endpointing/EndPointType.cs
@@ -1,0 +1,7 @@
+namespace WalletWasabi.IntegrationTests.BitcoinCore.Endpointing;
+
+public enum EndPointType
+{
+	P2p,
+	Rpc
+}
