+++ WalletWasabi.IntegrationTests/BundledApps/Binaries/.gitignore
@@ -1,0 +1,1 @@
+/temp/
