+++ WalletWasabi.IntegrationTests/WalletWasabi.IntegrationTests.csproj
@@ -24,12 +24,23 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\WalletWasabi\WalletWasabi.csproj" />
-		<ProjectReference Include="..\WalletWasabi.Tests\WalletWasabi.Tests.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<None Update="xunit.runner.json">
 			<CopyToOutputDirectory>Always</CopyToOutputDirectory>
+		</None>
+		<None Update="BundledApps\Binaries\linux-arm64\bitcoind">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+		<None Update="BundledApps\Binaries\linux-x64\bitcoind">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+		<None Update="BundledApps\Binaries\osx64\bitcoind">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+		<None Update="BundledApps\Binaries\win-x64\bitcoind.exe">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
 	</ItemGroup>
