--- WalletWasabi.Tests/BitcoinCore/Endpointing/EndPointStrategyType.cs
@@ -1,8 +1,0 @@
-namespace WalletWasabi.Tests.BitcoinCore.Endpointing;
-
-public enum EndPointStrategyType
-{
-	Default,
-	Custom,
-	Random
-}
