--- WalletWasabi.Tests/BitcoinCore/Endpointing/EndPointType.cs
@@ -1,7 +1,0 @@
-namespace WalletWasabi.Tests.BitcoinCore.Endpointing;
-
-public enum EndPointType
-{
-	P2p,
-	Rpc
-}
