--- WalletWasabi.Tests/BitcoinCore/Processes/BitcoindException.cs
@@ -1,12 +1,0 @@
-namespace WalletWasabi.Tests.BitcoinCore.Processes;
-
-public class BitcoindException : Exception
-{
-	public BitcoindException(string message) : base(message)
-	{
-	}
-
-	public BitcoindException(string message, Exception innerException) : base(message, innerException)
-	{
-	}
-}
