--- WalletWasabi.Tests/BundledApps/Binaries/.gitignore
@@ -1,1 +1,0 @@
-/temp/
